+++ oct-macros/src/lib.rs
@@ -15,7 +15,14 @@ extern crate self as oct_macros;
 
 use proc_macro::TokenStream;
 use quote::quote;
-use syn::{parse_macro_input, Data, DeriveInput};
+use syn::{
+	parse_macro_input,
+	parse2,
+	Data,
+	DeriveInput,
+	Type,
+};
+use syn::parse::Parse;
 
 mod derive_impl;
 
@@ -25,19 +32,33 @@ mod repr;
 use discriminants::Discriminants;
 use repr::Repr;
 
-#[proc_macro_derive(Decode)]
+#[proc_macro_derive(Decode, attributes(oct))]
 pub fn derive_decode(input: TokenStream) -> TokenStream {
 	let input = parse_macro_input!(input as DeriveInput);
 
 	let self_name = input.ident;
 
+	let mut error: Type = parse2(quote! { ::oct::error::GenericDecodeError }).unwrap();
+
+	for attr in &input.attrs {
+		if attr.meta.path().is_ident("oct") {
+			attr.parse_nested_meta(|meta| {
+				if meta.path.is_ident("decode_error") {
+					error = Parse::parse(meta.value()?)?;
+				}
+
+				Ok(())
+			}).unwrap();
+		}
+	}
+
 	let body = match input.data {
-		Data::Struct(data) => derive_impl::decode_struct(data),
+		Data::Struct(data) => derive_impl::decode_struct(data, error),
 
 		Data::Enum(data) => {
 			let repr = Repr::get(&input.attrs).unwrap_or_default();
 
-			derive_impl::decode_enum(data, repr)
+			derive_impl::decode_enum(data, repr, error)
 		}
 
 		Data::Union(_) => panic!("untagged union `{self_name}` cannot derive `oct::decode::Decode`"),
@@ -46,6 +67,7 @@ pub fn derive_decode(input: TokenStream) -> TokenStream {
 	let (impl_generics, ty_generics, where_clause) = input.generics.split_for_impl();
 
 	let output = quote! {
+		#[automatically_derived]
 		impl #impl_generics ::oct::decode::Decode for #self_name #ty_generics
 		#where_clause
 		{
@@ -56,19 +78,33 @@ pub fn derive_decode(input: TokenStream) -> TokenStream {
 	output.into()
 }
 
-#[proc_macro_derive(Encode)]
+#[proc_macro_derive(Encode, attributes(oct))]
 pub fn derive_encode(input: TokenStream) -> TokenStream {
 	let input = parse_macro_input!(input as DeriveInput);
 
 	let self_name = input.ident;
 
+	let mut error: Type = parse2(quote! { ::oct::error::GenericEncodeError }).unwrap();
+
+	for attr in &input.attrs {
+		if attr.meta.path().is_ident("oct") {
+			attr.parse_nested_meta(|meta| {
+				if meta.path.is_ident("encode_error") {
+					error = Parse::parse(meta.value()?)?;
+				}
+
+				Ok(())
+			}).unwrap();
+		}
+	}
+
 	let body = match input.data {
-		Data::Struct(data) => derive_impl::encode_struct(data),
+		Data::Struct(data) => derive_impl::encode_struct(data, error),
 
 		Data::Enum(data) => {
 			let repr = Repr::get(&input.attrs).unwrap_or_default();
 
-			derive_impl::encode_enum(data, repr)
+			derive_impl::encode_enum(data, repr, error)
 		}
 
 		Data::Union(_) => panic!("untagged union `{self_name}` cannot derive `oct::encode::Encode`"),
@@ -77,6 +113,7 @@ pub fn derive_encode(input: TokenStream) -> TokenStream {
 	let (impl_generics, ty_generics, where_clause) = input.generics.split_for_impl();
 
 	let output = quote! {
+		#[automatically_derived]
 		impl #impl_generics ::oct::encode::Encode for #self_name #ty_generics
 		#where_clause
 		{
@@ -108,6 +145,7 @@ pub fn derive_sized_encode(input: TokenStream) -> TokenStream {
 	let (impl_generics, ty_generics, where_clause) = input.generics.split_for_impl();
 
 	let output = quote! {
+		#[automatically_derived]
 		impl #impl_generics ::oct::encode::SizedEncode for #self_name #ty_generics
 		#where_clause
 		{